Move top-level App related components into their own subdirectory
